help
help i zmiany w funkcjach
